Report the provenance of duplicates better
When collecting all the generated resources in configaware.go, we have
an idea of where each one came from because it either has a config
file attached to it or not.

So we can report quite accurately where duplicate definitions arise --
they will either duplicate a resource in a YAML file (no config file
attached), the config file currently being processed, or a config file
previously processed.

To show the config file path as well as the --git-path in question,
I've combined them into a gitpath-plus-relative-path. For example, if
the original path is `staging`, but the .flux.yaml found is actually
in the directory above, it'll use `staging/../.flux.yaml` as the
location of the config file.