Detect duplicates in manifests parsed from bytes
When asssembling manifests from files or generators, there are some
checks for duplicate definitions. However, duplicates _within_ a
(multi-doc) file, or within a generated set of manifests, won't be
detected, since the code that parses a chunk of bytes (in
cluster/kubernetes/resource/load.go) doesn't check for duplicates.

So: add a test that duplicates within a generated set of manifests are
detected, and make it pass by checking when parsing bytes into
manifests.